requirements.txt
googletrans==2.2.0
requests==2.18.4
lxml==4.1.1
Flask==0.12.2
transitions==0.6.3
pygraphviz==1.3.1
python-telegram-bot==5.3.0